bugfix: shelf change on search results was not reflected when returning to main view except for none Needed to call update books for all cases not just none. Also moved Shelf Changer into components.
